Fix to .travis.yml to use Container Based testing
Enforcing sudo to false will tell Travis to make use of its Container
Based testing platform which is much faster. This is enabled by default
for new projects in Travis, but since Mongofill is older than the cutoff
date it continues to be tested on the legacy platform without this
manual change.