Move existing integration tests to separate Standalone namespace
In order to add separate replset tests for each class, it's easier
to re-organize the tests so that all replset tests are stored in
the same namespace. This way the tests can be grouped at the class
level rather than littered about with the standard integration
test suite which is run against a standalone connection.