Fixes MongoCollection to ensure read preference overrides cascade
If we are accessing a collection through a MongoDB object, we want
any read preference overrides applied to the MongoDB to cascade
to the MongoCollection object which, otherwise, would use the
read preference configured at the MongoClient level. This is
important for ReplicaSet configuration because we need to be able
to access the ReplicaSet configuration from the PRIMARy regardless
of the read preference set at the time the MongoClient is
initialized.