Fixes bug whereby wrong ReadPreference may be used in some situations
Because we need to talk to the PRIMARY during replSet initialization, we
were overriding this at the client level temporarily. Unfortunately this
causes an issue whereby some connections hang around with cursors that
are using the wrong read preference instead of the one originally
intended by the user. To fix we override it at the MongoDB level which
cascades to the MongoCollection object due to previous commit. This
ensures we read from the PRIMARY for the replica set intiialization
without affecting any other connections.